avdevice/dshow: implement capabilities API
This implements avdevice_capabilities_create and
avdevice_capabilities_free for the dshow device.

This enables configuration discovery of DirectShow devices through the
API, which is important for my use case. It enables making proper GUIs
presenting users with options, instead of asking them to discover a
dshow device's capabilities through the list_options option with an
FFmpeg tool, and listing what they want to configure in dumb text boxes.

Signed-off-by: Diederick Niehorster <dcnieho@gmail.com>